refactor: TypeScript型定義を適切な構造に整理
- custom.d.tsを削除し、型定義をsrc/types/に移動
- src/types/glsl.d.ts: GLSLファイルのモジュール定義
- src/types/three.d.ts: react-three/fiberの型拡張とカスタムマテリアル定義
- @ts-ignoreを削除し、適切な型安全性を確保
- Three.js実装の機能を維持しながら型エラーを完全解消
