+++ src/types/glsl.d.ts
@@ -1,0 +1,9 @@
+declare module '*.glsl?raw' {
+  const content: string;
+  export default content;
+}
+
+declare module '*.glsl' {
+  const content: string;
+  export default content;
+}
